Struct llvm::Arg

pub struct Arg;

An argument that is passed to a function.

Methods

impl Arg

fn add_attribute(&self, attr: Attribute)

Add the attribute given to this argument.

fn add_attributes(&self, attrs: &[Attribute])

Add all the attributes given to this argument.

fn has_attribute(&self, attr: Attribute) -> bool

Returns true if this argument has the attribute given.

fn has_attributes(&self, attrs: &[Attribute]) -> bool

Returns true if this argument has all the attributes given.

fn remove_attribute(&self, attr: Attribute)

Remove an attribute from this argument.