Struct llvm::Context

pub struct Context;

Contains all the LLVM entities - mainly modules.

Every single entity attached to it has its lifetime to enforce the rule that things from different contexts cannot interact and to preserve pointer safety.

Methods

impl Context

unsafe fn get_global() -> &'static Context

Get a reference to the global context.

This is marked as unsafe because this can result in undefined behaviour in a multithreaded context if they all use the same context.

fn new() -> CBox<Self>

Create a new context, which is owned by the callee block.