Struct llvm::Function

pub struct Function;

A function is a kind of value that can be called and contains blocks of code.

To get the value of each argument to a function, you can use the index operator. For example, &func[0] is the value that represents the first argument to the function.

Methods

impl Function

fn append<'a>(&'a self, name: &str) -> &'a BasicBlock

Add a basic block with the name given to the function and return it.

fn get_entry(&self) -> Option<&BasicBlock>

Returns the entry block of this function or None if there is none.

fn get_name(&self) -> &str

Returns the name of this function.

fn get_signature(&self) -> &FunctionType

Returns the function signature representing this function's signature.

fn add_attribute(&self, attr: Attribute)

Add the attribute given to this function.

fn add_attributes(&self, attrs: &[Attribute])

Add all the attributes given to this function.

fn has_attribute(&self, attr: Attribute) -> bool

Returns true if the attribute given is set in this function.

fn has_attributes(&self, attrs: &[Attribute]) -> bool

Returns true if all the attributes given is set in this function.

fn remove_attribute(&self, attr: Attribute)

Remove the attribute given from this function.