Struct llvm::Module

pub struct Module;

Represents a single translation unit of code.

This is attached to the lifetime of the context that constructs it, but is owned by the CSemiBox.

Methods

impl Module

fn new<'a>(name: &str, context: &'a Context) -> CSemiBox<'a, Module>

Create a new module in the context given with the name given.

The lifetime of the module will match the lifetime of the context you instance it in because the context contains it.

use llvm::*;
let context = Context::new();
let context = context.as_semi();
let module = Module::new("name", &context);
println!("{:?}", module)

fn add_global<'a>(&'a self, name: &str, ty: &'a Type) -> &'a Value

Add a global to the module with the given type and name.

fn add_global_constant<'a>(&'a self, name: &str, val: &'a Value) -> &'a Value

Add a constant global to the module with the given type, name and value.

fn get_global<'a>(&'a self, name: &str) -> Option<&'a Value>

Get the global with the name given, or None if no global with that name exists.

fn parse_bitcode<'a>(context: &'a Context, path: &str) -> Result<CSemiBox<'a, Module>, CBox<str>>

Parse this bitcode file into a module, or return an error string.

fn write_bitcode(&self, path: &str) -> IoResult<()>

Write this module's bitcode to the path given.

fn add_function<'a>(&'a self, name: &str, sig: &'a Type) -> &'a mut Function

Add a function to the module with the name given.

fn get_function<'a>(&'a self, name: &str) -> Option<&'a Function>

Returns the function with the name given if it exists.

fn get_type<'a>(&'a self, name: &str) -> Option<&'a Type>

Returns the type with the name given if it exists.

fn clone<'a>(&'a self) -> CSemiBox<'a, Module>

Clone this module.

fn optimize(&self, opt_level: usize, size_level: usize)

Optimize this module with the given optimization level and size level.

fn get_target(&self) -> &str

Returns the target data of this module represented as a string

fn set_target(&self, target: &str)

Set the target data of this module

fn verify(&self) -> Result<(), CBox<str>>

Verify that the module is safe to run.

fn compile(&self, path: &Path, opt_level: usize) -> IoResult<()>

Compile the module into an object file at the given location.

Note that this uses the LLVM tool llc to do this, which may or may not be installed on the user's machine.

fn link(&self, src: &Module) -> Result<(), CBox<str>>

Link a module into this module

fn link_destroy(&self, src: CSemiBox<Module>) -> Result<(), CBox<str>>

Link a module into this module