Struct llvm::ObjectFile

pub struct ObjectFile {
    // some fields omitted
}

An external object file that has been parsed by LLVLM

Methods

impl ObjectFile

fn read(path: &str) -> Result<ObjectFile, CBox<str>>

Attempt to parse the object file at the path given

fn symbols(&self) -> Symbols

Iterate through the symbols in this object fil