Struct llvm::Type

pub struct Type(_);

Defines how a value should be laid out in memory.

Methods

impl Type

fn get<'a, T>(context: &'a Context) -> &'a Type where T: Compile<'a>

Get the type given as an LLVM type descriptor in the context given.

fn is_sized(&self) -> bool

Returns true if the size of the type is known at compile-time.

This is equivalent to the type implementing Sized in Rust

fn is_function(&self) -> bool

Returns true if this type is a function.

This is equivalent to FunctionType::is.

fn is_struct(&self) -> bool

Returns true if this type is a struct.

This is equivalent to StructType::is.

fn is_void(&self) -> bool

Returns true if this type is void.

fn is_pointer(&self) -> bool

Returns true if this type is a pointer.

This is equivalent to PointerType::is.

fn is_integer(&self) -> bool

Returns true if this type is an integer.

fn is_float(&self) -> bool

Returns true if this type is any floating-point number.

fn get_size(&self, target: &TargetData) -> usize

Returns the size of the type in bytes.