Trait llvm::Compile

pub trait Compile<'a> {
    fn compile(self, context: &'a Context) -> &'a Value;
    fn get_type(context: &'a Context) -> &'a Type;
}

Implemented for any type that can be represented as a constant in IR.

Required Methods

fn compile(self, context: &'a Context) -> &'a Value

Compile this value in the context given.

fn get_type(context: &'a Context) -> &'a Type

Get the type descriptor for this type in the context given.