Trait llvm::ExecutionEngine

pub trait ExecutionEngine<'a, 'b> where LLVMExecutionEngineRef: From<&'b Self> {
    type Options: Copy;
    fn new(module: &'a Module, options: Self::Options) -> Result<Self, CBox<str>>;

    fn add_module(&'b self, module: &'a Module) { ... }
    fn remove_module(&'b self, module: &'a Module) -> &'a Module { ... }
    fn run_static_constructors(&'b self) { ... }
    fn run_static_destructors(&'b self) { ... }
    fn find_function(&'b self, name: &str) -> Option<&'a Function> { ... }
    fn run_function(&'b self, function: &'a Function, args: &[GenericValue<'a>]) -> GenericValue<'a> { ... }
    unsafe fn get_global<T>(&'b self, global: &'a Value) -> &'b T { ... }
    unsafe fn find_global<T>(&'b self, name: &str) -> Option<&'b T> { ... }
}

An abstract interface for implementation execution of LLVM modules.

This is designed to support both interpreter and just-in-time (JIT) compiler implementations.

Associated Types

type Options: Copy

The options given to this upon creation.

Required Methods

fn new(module: &'a Module, options: Self::Options) -> Result<Self, CBox<str>>

Create a new execution engine with the given Module and options, or return a description of the error.

Provided Methods

fn add_module(&'b self, module: &'a Module)

Add a module to the list of modules to interpret or compile.

fn remove_module(&'b self, module: &'a Module) -> &'a Module

Remove a module from the list of modules to interpret or compile.

fn run_static_constructors(&'b self)

Execute all of the static constructors for this program.

fn run_static_destructors(&'b self)

Execute all of the static destructors for this program.

fn find_function(&'b self, name: &str) -> Option<&'a Function>

Attempt to find a function with the name given, or None if there wasn't a function with that name.

fn run_function(&'b self, function: &'a Function, args: &[GenericValue<'a>]) -> GenericValue<'a>

Run function with the arguments given as `GenericValues, then return the result as one.

Note that if this engine is a JitEngine, it only supports a small fraction of combinations for the arguments and return value, so be warned.

To convert the arguments to GenericValues, you should use the GenericValueCast::to_generic method. To convert the return value from a GenericValue, you should use the GenericValueCast::from_generic method.

unsafe fn get_global<T>(&'b self, global: &'a Value) -> &'b T

Returns a pointer to the global value given.

This is marked as unsafe because the type cannot be guranteed to be the same as the type of the global value at this point.

unsafe fn find_global<T>(&'b self, name: &str) -> Option<&'b T>

Returns a pointer to the global value with the name given.

This is marked as unsafe because the type cannot be guranteed to be the same as the type of the global value at this point.