Struct llvm::Arg

pub struct Arg;

A value that represents an argument that has been given to a function.

Methods

impl Arg

fn add_attribute(&self, attr: Attribute)

Add an attribute to a function argument

fn add_attributes(&self, attrs: &[Attribute])

Add attributes to this function argument

fn has_attribute(&self, attr: Attribute) -> bool

Get the attributes set for a function argument

fn has_attributes(&self, attrs: &[Attribute]) -> bool

Check if this argument has all the attributes given

fn remove_attribute(&self, attr: Attribute)

Remove an attribute from a function argument