Struct llvm::BasicBlock

pub struct BasicBlock;

A container of instructions that execute sequentially.

Methods

impl BasicBlock

fn get_parent(&self) -> Option<&Value>

Return the enclosing method, or None if it is non-existant.

fn move_after(&self, other: &BasicBlock)

Move this basic block after the other basic block in its function.

fn move_before(&self, other: &BasicBlock)

Move this basic block before the other basic block in its function.

fn remove(&self)

Unlink from the containing function, but do not delete it.

unsafe fn delete(&self)

Delete this basic block.

This is unsafe because there should be no other reference to this, but this can't be guranteed.