Struct llvm::Builder

pub struct Builder;

This provides a uniform API for creating instructions and inserting them into a basic block.

Methods

impl Builder

fn new(context: &Context) -> CSemiBox<Builder>

Create a new builder in the context given.

fn position_at_end(&self, block: &BasicBlock)

Position the builder at the end of block.

fn build_ret_void(&self) -> &Value

Build an instruction that returns from the function with void.

fn build_ret(&self, value: &Value) -> &Value

Build an instruction that returns from the function with value.

fn build_array_alloca(&self, elem: &Type, size: &Value) -> &Value

Build an instruction that allocates an array with the element type elem and the size size.

The size of this array will be the size of elem times size.

fn build_alloca(&self, ty: &Type) -> &Value

Build an instruction that allocates a pointer to fit the size of ty then returns this pointer.

Make sure to call build_free with the pointer value when you're done with it, or you're gonna have a bad time.

fn build_free(&self, val: &Value) -> &Value

Build an instruction that frees the val, which MUST be a pointer that was returned from build_alloca.

fn build_store(&self, val: &Value, ptr: &Value) -> &Value

Build an instruction that store the value val in the pointer ptr.

fn build_br(&self, dest: &BasicBlock) -> &Value

Build an instruction that branches to the block dest.

fn build_cond_br(&self, cond: &Value, if_block: &BasicBlock, else_block: Option<&BasicBlock>) -> &Value

Build an instruction that branches to if_block if cond evaluates to true, and else_block otherwise.

fn build_call(&self, func: &Function, args: &[&Value]) -> &Value

Build an instruction that calls the function func with the arguments args.

This will return the return value of the function.

fn build_tail_call(&self, func: &Function, args: &[&Value]) -> &Value

Build an instruction that calls the function func with the arguments args.

This will return the return value of the function.

fn build_select(&self, cond: &Value, true_val: &Value, false_val: &Value) -> &Value

Build an instruction that yields to true_val if cond is equal to 1, and false_val otherwise.

fn build_bit_cast(&self, value: &Value, dest: &Type) -> &Value

Build an instruction that casts a value into a certain type.

fn build_insert_value(&self, agg: &Value, elem: &Value, index: usize) -> &Value

Build an instruction that inserts a value into an aggregate type.

fn build_gep(&self, pointer: &Value, indices: &[&Value]) -> &Value

Build an instruction that computes the address of a subelement of an aggregate data structure.

Basically type-safe pointer arithmetic.

fn build_switch(&self, value: &Value, default: &BasicBlock, cases: &[(&Value, &BasicBlock)]) -> &Value

Build an instruction that runs whichever block matches the value, or default if none of them matched it.

fn build_load(&self, value: &Value) -> &Value

fn build_neg(&self, value: &Value) -> &Value

fn build_not(&self, value: &Value) -> &Value

fn build_add(&self, left: &Value, right: &Value) -> &Value

fn build_sub(&self, left: &Value, right: &Value) -> &Value

fn build_mul(&self, left: &Value, right: &Value) -> &Value

fn build_div(&self, left: &Value, right: &Value) -> &Value

fn build_shl(&self, left: &Value, right: &Value) -> &Value

fn build_ashr(&self, left: &Value, right: &Value) -> &Value

fn build_and(&self, left: &Value, right: &Value) -> &Value

fn build_or(&self, left: &Value, right: &Value) -> &Value

fn build_cmp(&self, a: &Value, b: &Value, pred: Predicate) -> &Value

Build an instruction to compare two values with the predicate