Struct llvm::CSemiBox

pub struct CSemiBox<'a, D> where D: DisposeRef + 'a + ?Sized {
    // some fields omitted
}

A wrapper for pointers made by C that are now partially owned in Rust.

This is necessary to allow owned and borrowed representations of C types to be represented by the same type as they are in C with little overhead

Methods

impl<'a, D> CSemiBox<'a, D> where D: DisposeRef + 'a + ?Sized

fn new(ptr: *mut D::RefTo) -> CSemiBox<'a, D>

Wrap the pointer in a CSemiBox

unsafe fn as_ptr(&self) -> *mut D::RefTo

Returns the internal pointer

unsafe fn unwrap(self) -> *mut D::RefTo

Returns the internal pointer