Struct llvm::ObjectFile

pub struct ObjectFile {
    // some fields omitted
}

An external object file that has been parsed by LLVM.

Methods

impl ObjectFile

fn read(path: &str) -> Result<ObjectFile, CBox<str>>

Parse the object file at the path given, or return an error string if an error occurs.

fn symbols(&self) -> Symbols

Iterate through the symbols in this object file.