Struct llvm::Target

pub struct Target;

Methods

impl Target

fn get_name(&self) -> &str

Returns the name of this target.

fn get_description(&self) -> &str

Returns the description of this target.

fn has_asm_backend(&self) -> bool

Returns true if this target has an assembly generation backend in LLVM.

fn has_jit(&self) -> bool

Returns true if this target supports JIT compilation.

fn has_target_machine(&self) -> bool

Returns true if this target has a target machine.