Struct llvm::Value

pub struct Value;

A typed value that can be used as an operand in instructions.

Methods

impl Value

fn new_struct<'a>(context: &'a Context, vals: &[&'a Value], packed: bool) -> &'a Value

Create a new constant struct from the values given.

fn new_vector<'a>(vals: &[&'a Value]) -> &'a Value

Create a new constant vector from the values given.

fn new_string<'a>(context: &'a Context, text: &str, rust_style: bool) -> &'a Value

Create a new constant C string from the text given.

fn new_undef<'a>(ty: &'a Type) -> &'a Value

Create a new constant undefined value of the given type.

fn get_name(&self) -> Option<&str>

Returns the name of this value, or None if it lacks a name

fn set_name(&self, name: &str)

Sets the name of this value

fn get_type(&self) -> &Type

Returns the type of this value